Add files via upload
This commit provides libhackrf.  I did *not* write this code; it was created by the original authors.  The only change that I made to the hackrf.c file is to fix the issue where the system would sometimes get stuck in transmit mode.  I submitted the diff to the appropriate hackrf developers, and they incorporated my change into the baseline code.