This commit incorporates the changes listed below:
1. Responsibility of instantiation of the frequency scanner has been moved
from the diag UI to the mainline code.

2. The user now has to configure the frequency scanner before starting a
scan if new frequency parameters are desired.

3. The user can enable/disable the scanner without having to change the
scanning parameters.  This allows the user to stop scanning (which
provides pause functionality), and restart the scan with the result that
the scan will continue from the last frequency for which the scanner was
stopped.

4. The frequency scanner exists for the whole lifetime of the program.

This should make things easier when if I decide to absorbe scanner
functionality into the Radio object. For now, I'll let it be a separate
Radio object application.